Better team compositions
I dumped my old algorithm to assign teams based on the 'next team for a level'.  This was resulting in teams that did not have the correct number of people on it.  I attempted to try a round of 'fix up' on this and it was a disaster.  The new algorithm, documented in the readme, works far better.